Bare minimum needed to make this compile. Oh, and change tmpname to a char.

+++ 8086/c86/scan.c
@@ -102,7 +102,7 @@ mainp1( argc, argv ) int argc; char *argv[]; {  /* control multiple files */
 
 	/* dimension table initialization */
 
-	dimtab[NULL] = 0;
+	dimtab[(int)NULL] = 0;
 	dimtab[CHAR] = SZCHAR;
 	dimtab[INT] = SZINT;
 	dimtab[FLOAT] = SZFLOAT;
